// src/interfaces.tsx
import { MouseEventHandler, KeyboardEventHandler } from 'react';

export interface ICharterForm {
  birth: string;
  death: string;
  gender: string;
  hair: string;
  height: string;
  name: string;
  race: string;
  realm: string;
  spouse: string;
  wikiUrl: string;
  file: File[];
}

export interface ICharacter {
  birth: string;
  death: string;
  gender: string;
  hair: string;
  height: string;
  name: string;
  race: string;
  realm: string;
  spouse: string;
  wikiUrl: string;
  _id: string;
  image: string;
}
export interface ICharacterNew {
  position: number;
  birth: string;
  death: string;
  gender: string;
  hair: string;
  height: string;
  name: string;
  race: string;
  realm: string;
  spouse: string;
  wikiUrl: string;
  _id: string;
  image: string;
}
export interface ICharacterMod {
  birth: string;
  death: string;
  gender: string;
  hair: string;
  height: string;
  name: string;
  race: string;
  realm: string;
  spouse: string;
  wikiUrl: string;
  _id: string;
  image: string;
  cbCloseModal: MouseEventHandler<HTMLElement>;
}

export interface ISearch {
  cbSearch(data: ICharacter[]): void;
}

export interface IStateAction {
  type: 'addValue';
  payload: string;
}
export interface ISearchState {
  search: string;
}